test(#7778): move text block walker tests into their own class
